src/main/java/awsGroup/awsPractice/com/sikdan/webController/HelloController.java
package awsGroup.awsPractice.com.sikdan.webController;

import awsGroup.awsPractice.com.sikdan.webController.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class HelloController {
    @GetMapping("/hello")
    public String hello() {
        return "hello";
    }

    @GetMapping("/hello/dto")
    public HelloResponseDto saveDTO(@RequestParam("name") String name, @RequestParam("amount") int amount) {
        HelloResponseDto response = new HelloResponseDto(name, amount);

        return response;
    }
}
